requirements.txt
discord.py
aiosqlite
gradio_client
ImageHash
Pillow
python-dotenv
